test edge case cancel behavior in edge case where multiple orders exist where cancelCol (for cancelReplace) are set by same id.
The correct behavior, for performance, is to terminate the scan early for orders to cancel when a previous cancelReplace is found.
Although this could  result in not all cancellable orders being deleted, this is a tradeoff for better performance.
Also since the previous cancelReplace should have deleted all pervious orders it should not impact the active order set under
ordinary conditions.